agents/content-wizard/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
huggingface_hub
edge-tts
asyncionest_asyncio

